Normalize whitespace
Most of the source code in this extension uses tabs,
but there were 26 lines using spaces as indentation.
This probably snuck in during a PR, but it makes
portions of the source difficult to read on github.com
and probably elsewhere.

Normalize those 26 lines to match the rest of the file.